fix another two gcc warnings
git-svn-id: svn://svn.powerdns.com/pdns/trunk/pdns@2562 d19b8d6e-7fed-0310-83ef-9ca221ded41b

+++ pdns/dynlistener.cc
@@ -265,7 +265,7 @@ void DynListener::sendLine(const string &l)
     if(!line.empty() && line[line.length()-1]!='\n')
       line.append("\n");
     line.append("\n");
-    if(write(1,line.c_str(),line.length()) != line.length())
+    if((unsigned int)write(1,line.c_str(),line.length()) != line.length())
       L<<Logger::Error<<"Error sending data to console: "<<stringerror()<<endl;
       
   }
